Change string_view to char array for STARTPOS_FEN
Replaced the string_view declaration of STARTPOS_FEN with a character array in Config.h. This was done for the purpose of using STARTPOS_FEN without invoking .data() in Board.cpp function. It improves code readability and efficiency. Likewise, in Board.h, the explicit false initialization of chess960 was removed as bool is false by default. It streamlines the code and avoids redundancy.